requirements.txt
albucore==0.0.23
albumentations==2.0.5
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.13.1
flatbuffers==25.2.10
fonttools==4.55.2
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.6.1
iopath==0.1.10
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.16.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==2.1.1
omegaconf==2.3.0
onnx==1.17.0
onnxruntime-gpu==1.21.0
onnxslim==0.1.49
opencv-contrib-python==4.10.0.84
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
pickleshare==0.7.5
pillow==11.0.0
platformdirs==4.3.7
portalocker==3.0.0
prompt_toolkit==3.0.50
protobuf==5.29.1
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==18.1.0
pycocotools==2.0.8
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydicom==3.0.1
Pygments==2.18.0
pyparsing==3.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scipy==1.15.2
seaborn==0.13.2
segment-anything==1.0
setuptools==75.1.0
simsimd==6.2.1
six==1.17.0
smmap==5.0.1
stack-data==0.6.3
streamlit==1.40.2
streamlit-drawable-canvas==0.9.3
stringzilla==3.12.1
sympy==1.13.1
tenacity==9.0.0
toml==0.10.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.98
ultralytics-thop==2.0.14
urllib3==2.2.3
watchdog==6.0.0
wcwidth==0.2.13
wheel==0.44.0
zipp==3.21.0

# 只在 Windows 安裝 pywin32
pywin32==308; sys_platform == "win32"

# 安裝 SAM_2（來自 GitHub）
git+https://github.com/111360148/EasyMedVision.git@521d3101b7506dd869763edfc909493c5a49c136#egg=SAM_2